Module Ocastore_eio

Disk and network backends for Ocastore databases.

Disk

val create : _ Eio.Path.t -> 'a Ocastore.schema -> 'a -> 'a Ocastore.db

create path schema x initializes a fresh database on disk at location path, with the initial value x described by schema.

val open_rw : _ Eio.Path.t -> 'a Ocastore.schema -> 'a Ocastore.db option

open_rw path schema attempts to open the database at location path on disk, which stores an 'a value described by the schema. Returns Some db on success, or None if the database doesn't exists at that location.

Example, to try to open an existing database or create it otherwise:

let db_path = Path.(Stdenv.cwd env / "mydb") in
let db = match Ocastore_eio.open_rw db_path typerep_of_t with
  | Some db -> db
  | None -> Ocastore_eio.create db_path typerep_of_t initial_value
in
(* ... *)

Network

val serve : sw:Eio.Switch.t -> _ Eio.Net.t -> Eio.Net.Sockaddr.stream -> _ Ocastore.db -> _

serve ~sw net addr db runs a server exposed at address addr to serve the database db over the network. The server never returns, unless the switch sw is toggled off.

Example:

let db = (* create_in_memory or create/open_rw or ... *) in
Switch.run @@ fun sw ->
Ocastore_eio.serve ~sw (Stdenv.net env) (`Tcp (Net.Ipaddr.V4.any, 1234)) db

val connect : sw:Eio.Switch.t -> _ Eio.Net.t -> Eio.Net.Sockaddr.stream -> 'a Ocastore.schema -> 'a Ocastore.db

connect ~sw net addr schema returns a database connected to the server at address addr which serves a database of the same schema. The schema must be exactly the same one used by serve.

Example:

Switch.run @@ fun sw ->
let db =
  Ocastore_eio.connect ~sw (Stdenv.net env) (`Tcp (Net.Ipaddr.V4.any, 1234)) schema
in
(* ... query and update [db] as usual ... *)